modificamos modelo message

+++ models/message.go
@@ -3,5 +3,5 @@ package models
 // Message model
 type Message struct {
 	Message string `json:"message"`
-	Code    uint8  `json:"code"`
+	Code    int    `json:"code"`
 }
